+++ libs/async_mpi/README.rst
@@ -7,7 +7,7 @@
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 ===
-mpi
+async_mpi
 ===
 
 This library is part of HPX.
