+++ libs/async_mpi/include/hpx/async_mpi/mpi_executor.hpp
@@ -9,7 +9,7 @@
 #include <hpx/config.hpp>
 #include <hpx/basic_execution/execution.hpp>
 #include <hpx/execution/traits/is_executor.hpp>
-#include <hpx/mpi/mpi_future.hpp>
+#include <hpx/async_mpi/mpi_future.hpp>
 #include <hpx/mpi_base.hpp>
 
 #include <type_traits>
